requirements/dev.txt
-r common.txt

drf-yasg==1.20.0
flake8==3.8.4
codecov==2.1.13
python-dotenv==0.20.0
config==0.5.1
pre-commit==2.20.0
boto3==1.26.161
django-storages==1.13.2
